// src/app/models/app-models.ts
/* eslint-disable @typescript-eslint/naming-convention */
export class Student {
    id: number;
    name: string;
    age: string;
    address: string;
}

export class Beer {
    id: number;
    brand: string;
    description: string;
    price: number;
}

export class User {
    id: number;
    fist_name?: string;
    last_name?: string;
    name?: string;
    email?: string;
    password?: string;
}